+++ src/main/java/codes/biscuit/skyblockaddons/utils/objects/ReturnValue.java
@@ -1,4 +1,4 @@
-package codes.biscuit.skyblockaddons.utils;
+package codes.biscuit.skyblockaddons.utils.objects;
 
 public class ReturnValue<R> {
